+++ ui/mods/com.pa.quitch.gwaioverhaul/gw_play/referee_config.js
@@ -135,7 +135,8 @@ define([
         inventory,
         playerTag,
         refereeCoop.getOrderedSubcommanders(inventory, game).length,
-        battleRng
+        battleRng,
+        { ffa: !_.isEmpty(ai.foes) }
       );
     }